doc: reformat group definitions, disable group graph
